requirements.txt
aioredis==1.3.1
asgi-redis==1.4.3
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
channels==3.0.3
channels-redis==3.3.0
charset-normalizer==2.0.4
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
dj-database-url==0.5.0
Django==3.2.2
django-ckeditor==6.1.0
django-friendship==1.9.1
django-heroku==0.3.1
django-js-asset==1.2.2
django-storages==1.11.1
dropbox==11.18.0
google-api-core==2.0.1
google-auth==2.0.2
google-cloud-core==2.0.0
google-cloud-storage==1.42.0
google-crc32c==1.1.2
google-resumable-media==2.0.2
googleapis-common-protos==1.53.0
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
msgpack==1.0.2
msgpack-python==0.5.6
Pillow==8.2.0
ply==3.11
protobuf==3.17.3
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pytils==0.3
pytz==2021.1
redis==2.10.6
requests==2.26.0
rsa==4.7.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
stone==3.2.1
Twisted==21.7.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typing-extensions==3.10.0.2
urllib3==1.26.6
whitenoise==5.2.0
zope.interface==5.4.0
